+++ app/src/main/java/com/tirexmurina/testapp/presentation/adapters/CategoriesListAdapter.kt
@@ -1,9 +1,8 @@
-package com.tirexmurina.testapp.presentation
+package com.tirexmurina.testapp.presentation.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.tirexmurina.testapp.databinding.ItemBannerListBinding
 import com.tirexmurina.testapp.databinding.ItemCategoryInactiveBinding
 
 class CategoriesListAdapter () : RecyclerView.Adapter<CategoriesItemViewHolder>() {
